refactor: move recovery markers into recovery module
